+++ espr/src/ir/type_decl.rs
@@ -94,6 +94,41 @@ impl Legalize for TypeDecl {
                     .collect::<Result<Vec<_>, _>>()?;
                 TypeDecl::Select(Select { id, types })
             }
+            Type::Set { base, bound } => {
+                let base = TypeRef::legalize(ns, ss, scope, base.as_ref())?;
+                let bound = if let Some(bound) = bound {
+                    Some(Legalize::legalize(ns, ss, scope, bound)?)
+                } else {
+                    None
+                };
+                TypeDecl::Rename(Rename {
+                    id,
+                    ty: TypeRef::Set {
+                        base: Box::new(base),
+                        bound,
+                    },
+                })
+            }
+            Type::List {
+                base,
+                bound,
+                unique,
+            } => {
+                let base = TypeRef::legalize(ns, ss, scope, base.as_ref())?;
+                let bound = if let Some(bound) = bound {
+                    Some(Legalize::legalize(ns, ss, scope, bound)?)
+                } else {
+                    None
+                };
+                TypeDecl::Rename(Rename {
+                    id,
+                    ty: TypeRef::List {
+                        base: Box::new(base),
+                        bound,
+                        unique: *unique,
+                    },
+                })
+            }
             _ => panic!(),
         })
     }